Assignment 3

+++ routes/dishRouter.js
@@ -21,27 +21,39 @@ dishRouter.route('/')
     .catch((err)=> next(err));
 })
 .post(authenticate.verifyUser, (req,res,next)=> {
-    Dishes.create(req.body)
-    .then((dish)=>{
-        console.log('Dish Created ', dish);
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(dish);
-    }, (err)=> next(err))
-    .catch((err)=>next(err));
+    if(authenticate.verifyAdmin(req.user.admin)){
+        Dishes.create(req.body)
+        .then((dish)=>{
+            console.log('Dish Created ', dish);
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(dish);
+        }, (err)=> next(err))
+        .catch((err)=>next(err));
+    } else {
+        err = new Error('You are not authorized');
+        err.status = 403;
+        return next(err);
+    }
 })
 .put(authenticate.verifyUser, (req,res,next)=> {
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes');
 })
-.delete(authenticate.verifyUser, (req,res,next)=> {
-    Dishes.remove({})
-    .then((resp)=>{
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err)=> next(err))
-    .catch((err)=>next(err));
+.delete(authenticate.verifyUser,(req,res,next)=> {
+    if( authenticate.verifyAdmin(req.user.admin)){
+        Dishes.remove({})
+        .then((resp)=>{
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err)=> next(err))
+        .catch((err)=>next(err));
+    } else {
+        err = new Error('You are not authorized');
+        err.status = 403;
+        return next(err);
+    }
 });
 // for Id
 dishRouter.route('/:dishId')
@@ -60,28 +72,40 @@ dishRouter.route('/:dishId')
     res.end('POST operation not supported on /dishes/ ' + req.params.dishId);
 })
 .put(authenticate.verifyUser, (req, res, next)=> {
-    Dishes.findByIdAndUpdate(req.params.dishId, {
-        $set: req.body
-    },{new: true})
-    .then((dish)=>{
-        Dishes.findById(dish._id)
-            .populate('comments.author')
-            .then((dish)=>{
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);
-            },(err)=>next(err))
-    }, (err)=> next(err))
-    .catch((err)=>next(err));
+    if(authenticate.verifyAdmin(req.user.admin)){
+        Dishes.findByIdAndUpdate(req.params.dishId, {
+            $set: req.body
+        },{new: true})
+        .then((dish)=>{
+            Dishes.findById(dish._id)
+                .populate('comments.author')
+                .then((dish)=>{
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(dish);
+                },(err)=>next(err))
+        }, (err)=> next(err))
+        .catch((err)=>next(err));
+    } else {
+        err = new Error('You are not authorized');
+        err.status = 403;
+        return next(err);
+    }
 })
 .delete(authenticate.verifyUser, (req, res, next)=> {
-    Dishes.findByIdAndRemove(req.params.dishId)
-    .then((resp)=>{
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err)=> next(err))
-    .catch((err)=>next(err));
+    if(authenticate.verifyAdmin(req.user.admin)) {
+        Dishes.findByIdAndRemove(req.params.dishId)
+        .then((resp)=>{
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err)=> next(err))
+        .catch((err)=>next(err));
+    } else {
+        err = new Error('You are not authorized');
+        err.status = 403;
+        return next(err);
+    }
 });
 
 //for comment
@@ -97,7 +121,7 @@ dishRouter.route('/:dishId/comments')
             res.json(dish.comments);
         }else{
             err = new Error('Dish '+ req.params.dishId + ' not found');
-            er.status = 404;
+            err.status = 404;
             return next(err);
         }
     }, (err)=> next(err))
@@ -107,7 +131,7 @@ dishRouter.route('/:dishId/comments')
     Dishes.findById(req.params.dishId)
     .then((dish)=>{
         if(dish != null) {
-            req.body.author = req.user._id; //
+            req.body.author = req.user._id; // this
             dish.comments.push(req.body);
             dish.save()
             .then((dish)=> {
@@ -117,7 +141,7 @@ dishRouter.route('/:dishId/comments')
             },(err)=> next(err))
         }else{
             err = new Error('Dish '+ req.params.dishId + ' not found');
-            er.status = 404;
+            err.status = 404;
             return next(err);
         }
     }, (err)=> next(err))
@@ -133,7 +157,11 @@ dishRouter.route('/:dishId/comments')
     .then((dish)=>{
         if(dish != null) {
             for (let i = (dish.comments.length -1); i>=0; i--){
-                dish.comments.id(dish.comments[i]._id).remove();
+                if(JSON.stringify(dish.comments.id(dish.comments[i]._id).author) === 
+                    JSON.stringify(req.user._id)) {
+                    
+                        dish.comments.id(dish.comments[i]._id).remove()
+                }
             }
             dish.save()
             .then((dish)=>{
@@ -162,11 +190,11 @@ dishRouter.route('/:dishId/comments/:commentId')
         }
         else if(dish == null){
             err = new Error('Dish '+ req.params.dishId + ' not found');
-            er.status = 404;
+            err.status = 404;
             return next(err);
         } else {
             err = new Error('Comment '+ req.params.commentId + ' not found');
-            er.status = 404;
+            err.status = 404;
             return next(err);
         }    
     }, (err)=> next(err))
@@ -181,23 +209,29 @@ dishRouter.route('/:dishId/comments/:commentId')
     Dishes.findById(req.params.dishId)
     .then((dish)=>{
         if(dish != null && dish.comments.id(req.params.commentId) != null) {
-            if(req.body.rating){
-                dish.comments.id(req.params.commentId).rating = req.body.rating;
-            }
-            if(req.body.comment){
-                dish.comments.id(req.params.commentId).comment = req.body.comment;
-            }
-            dish.save()
-            .then((dish)=>{
-                Dishes.findById(dish._id)
-                .populate('comments.author')
+            if(JSON.stringify(dish.comments.id(req.params.commentId).author) === 
+                    JSON.stringify(req.user._id)){
+                if(req.body.rating){
+                    dish.comments.id(req.params.commentId).rating = req.body.rating;
+                }
+                if(req.body.comment){
+                    dish.comments.id(req.params.commentId).comment = req.body.comment;
+                }
+                dish.save()
                 .then((dish)=>{
-                    res.statusCode =200;
-                    res.setHeader('Content-Type', 'application/json');
-                    res.json(dish);
-                },(err)=>next(err))
-            }, (err) => next(err));
-            
+                    Dishes.findById(dish._id)
+                    .populate('comments.author')
+                    .then((dish)=>{
+                        res.statusCode =200;
+                        res.setHeader('Content-Type', 'application/json');
+                        res.json(dish);
+                    },(err)=>next(err))
+                }, (err) => next(err));
+            } else {
+                err = new Error('You are not authorized to update this comment');
+                err.status = 403;
+                return next(err);
+            }      
         }
         else if(dish == null){
             err = new Error('Dish '+ req.params.dishId + ' not found');
@@ -205,7 +239,7 @@ dishRouter.route('/:dishId/comments/:commentId')
             return next(err);
         } else {
             err = new Error('Comment '+ req.params.commentId + ' not found');
-            er.status = 404;
+            err.status = 404;
             return next(err);
         }    
     }, (err)=> next(err))
@@ -215,25 +249,34 @@ dishRouter.route('/:dishId/comments/:commentId')
     Dishes.findById(req.params.dishId)
     .then((dish)=> {
         if(dish != null && dish.comments.id(req.params.commentId)!=null) {
-            dish.comments.id(req.params.commentId).remove();
-            dish.save()
-            .then((dish)=>{
-                Dishes.findById(dish._id)
-                .populate('comments.author')
+            if(JSON.stringify(dish.comments.id(req.params.commentId).author) === 
+                    JSON.stringify(req.user._id)){
+
+                dish.comments.id(req.params.commentId).remove();
+                dish.save()
                 .then((dish)=>{
-                    res.statusCode = 200;
-                    res.setHeader('Content-Type', 'application/json');
-                    res.json(dish);
-                },(err)=>next(err))
-            }, (err)=> next(err));
+                    Dishes.findById(dish._id)
+                    .populate('comments.author')
+                    .then((dish)=>{
+                        res.statusCode = 200;
+                        res.setHeader('Content-Type', 'application/json');
+                        res.json(dish);
+                    },(err)=>next(err))
+                }, (err)=> next(err));
+            } else {
+                err = new Error('You are not authorized to delete this comment');
+                err.status = 403;
+                return next(err);
+            }
+            
         } 
         else if(dish == null){
             err = new Error('Dish '+ req.params.dishId + ' not found');
-            er.status = 404;
+            err.status = 403;
             return next(err);
         } else {
             err = new Error('Comment '+ req.params.commentId + ' not found');
-            er.status = 404;
+            err.status = 404;
             return next(err);
         }    
     })
