+++ routes/leaderRouter.js
@@ -20,26 +20,38 @@ leaderRouter.route('/')
     .catch((err)=> next(err));
 })
 .post(authenticate.verifyUser,(req,res,next)=>{
-    Leaders.create(req.body)
-    .then((leader)=>{
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(leader);
-    }, (err)=>next(err))
-    .catch((err)=>next(err));
+    if(authenticate.verifyAdmin(req.user.admin)) {
+        Leaders.create(req.body)
+        .then((leader)=>{
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(leader);
+        }, (err)=>next(err))
+        .catch((err)=>next(err));
+    } else {
+        err = new Error('You are not authorized');
+        err.status = 403;
+        return next(err);
+    }
 })
 .put(authenticate.verifyUser,(req,res,next)=> {
     res.statusCode = 403;
     res.end('PUT operation not supported on /leaders');
 })
 .delete(authenticate.verifyUser,(req,res,next)=>{
-    Leaders.remove({})
-    .then((resp)=>{
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err)=>next(err))
-    .catch((err)=>next(err));
+    if(authenticate.verifyAdmin(req.user.admin)){
+        Leaders.remove({})
+        .then((resp)=>{
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err)=>next(err))
+        .catch((err)=>next(err));
+    } else {
+        err = new Error('You are not authorized');
+        err.status = 403;
+        return next(err);
+    }
 });
 //for /:promoId
 leaderRouter.route('/:leaderId')
@@ -57,24 +69,36 @@ leaderRouter.route('/:leaderId')
     res.end('POST operation not supported on /leaders/ ' + req.params.leaderId);
 })
 .put(authenticate.verifyUser,(req,res,next)=> {
-    Leaders.findByIdAndUpdate(req.params.leaderId,{
-        $set: req.body
-    },{new: true})    
-    .then((leader)=>{
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(leader);
-    },(err)=>next(err))
-    .catch((err)=>next(err));
+    if(authenticate.verifyAdmin(req.user.admin)) {
+        Leaders.findByIdAndUpdate(req.params.leaderId,{
+            $set: req.body
+        },{new: true})    
+        .then((leader)=>{
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(leader);
+        },(err)=>next(err))
+        .catch((err)=>next(err));
+    } else {
+        err = new Error('You are not authorized');
+        err.status = 403;
+        return next(err);
+    }
 })
 .delete(authenticate.verifyUser,(req,res,next)=> {
-    Leaders.findByIdAndRemove(req.params.leaderId)
-    .then((resp)=>{
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err)=>next(err))
-    .catch((err)=>next(err));
+    if(authenticate.verifyAdmin(req.user.admin)){
+        Leaders.findByIdAndRemove(req.params.leaderId)
+        .then((resp)=>{
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err)=>next(err))
+        .catch((err)=>next(err));
+    } else {
+        err = new Error('You are not authorized');
+        err.status = 403;
+        return next(err);
+    }
 });
